requirements.txt
aiosmtpd==1.4.2
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.2
atpublic==3.1.1
attrs==21.2.0
autopep8==1.7.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
distro-info===1.1build1
Django==4.1.2
django-crispy-forms==1.14.0
django-filter==22.1
django-grappelli==3.0.3
django-jazzmin==2.5.0
django-oauth-toolkit==2.2.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
docutils==0.19
duplicity==0.8.21
fasteners==0.14.1
Flask==2.2.2
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
html5lib==1.1
httpie==3.2.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipython==7.31.1
itsdangerous==2.1.2
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
jwcrypto==1.4.2
keyring==23.5.0
Kivy==2.1.0
Kivy-Garden==0.1.5
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==0.0.0
multidict==6.0.2
mysql-connector-python==8.0.30
mysqlclient==2.1.1
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
ply==3.11
prompt-toolkit==3.0.28
protobuf==3.12.4
ptyprocess==0.7.0
PyAutoGUI==0.9.53
pycairo==1.20.1
pycodestyle==2.9.1
pycups==2.0.1
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
PySocks==1.7.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python3-xlib==0.15
pythran==0.10.0
pytweening==1.0.4
pytz==2022.1
pywhatkit==5.4
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.10.1
rich==12.6.0
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stripe==2.63.0
sympy==1.9
systemd-python==234
toml==0.10.2
traitlets==5.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wikipedia==1.4.0
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
